+++ database/migrations/2020_07_19_154942_create_farmer_sales_view.php
@@ -18,10 +18,10 @@ class CreateFarmerSalesView extends Migration
         SELECT
             posts.user_id AS user_id,
             posts.crop_name AS crop_name,
-            SUM(posts.fixed_quantity) AS totalfixedqty,
-            SUM(posts.crop_quantity) AS totalavailableqty,
-            SUM(posts.kilogram_sold) AS totalkgsold,
-            AVG(ROUND(posts.percentage_sold_before_harvest,
+            SUM(posts.fixed_quantity::decimal) AS totalfixedqty,
+            SUM(posts.crop_quantity::decimal) AS totalavailableqty,
+            SUM(posts.kilogram_sold::decimal) AS totalkgsold,
+            AVG(ROUND(posts.percentage_sold_before_harvest::decimal,
                     1)) AS totalpercentage
         FROM
             posts
