+++ database/migrations/2020_07_19_160911_create_farmer_totalqtyandcropprice_view.php
@@ -17,8 +17,8 @@ class CreateFarmerTotalqtyandcroppriceView extends Migration
         CREATE VIEW farmer_totalqtyandcropprice AS
             SELECT
                 posts.user_id AS user_id,
-                SUM(posts.crop_quantity) AS sumcropqty,
-                AVG(posts.crop_price) AS avgcropprice,
+                SUM(posts.crop_quantity::decimal) AS sumcropqty,
+                AVG(posts.crop_price::decimal) AS avgcropprice,
                 posts.crop_name AS crop_name
             FROM
                 posts
