+++ database/migrations/2020_07_21_153757_create_chart_sales2_view.php
@@ -17,10 +17,10 @@ class CreateChartSales2View extends Migration
     CREATE VIEW chart_sales2 AS
     SELECT
         posts.crop_name AS crop_name,
-        SUM(posts.fixed_quantity) AS totalfixedqty,
-        SUM(posts.crop_quantity) AS totalavailableqty,
-        SUM(posts.kilogram_sold) AS totalkgsold,
-        AVG(ROUND(posts.percentage_sold_before_harvest,
+        SUM(posts.fixed_quantity::decimal) AS totalfixedqty,
+        SUM(posts.crop_quantity::decimal) AS totalavailableqty,
+        SUM(posts.kilogram_sold::decimal) AS totalkgsold,
+        AVG(ROUND(posts.percentage_sold_before_harvest::decimal,
                 1)) AS totalpercentage,
         created_at
     FROM
